requirements.txt
google.generativeai
python-dotenv
ipython
requests
